+++ src/tetra_rp/cli/main.py
@@ -11,6 +11,7 @@ from .commands import (
     build,
     # resource,
     deploy,
+    undeploy,
 )
 
 
@@ -37,7 +38,6 @@ app.command("init")(init.init_command)
 app.command("run")(run.run_command)
 app.command("build")(build.build_command)
 # app.command("report")(resource.report_command)
-# app.command("clean")(resource.clean_command)
 
 
 # command: flash deploy
@@ -57,6 +57,12 @@ deploy_app.command("new")(deploy.new_command)
 
 # command: flash deploy *
 app.add_typer(deploy_app, name="deploy")
+
+
+# command: flash undeploy
+# Note: Using a simple command instead of sub-app to allow both "undeploy list" and "undeploy <name>"
+# The undeploy_command internally handles the "list" case
+app.command("undeploy")(undeploy.undeploy_command)
 
 
 @app.callback(invoke_without_command=True)