+++ CHANGELOG.md
@@ -1,5 +1,36 @@
 # Changelog
 
+## [0.18.0](https://github.com/runpod/tetra-rp/compare/v0.17.1...v0.18.0) (2025-11-28)
+
+
+### Features
+
+* **cli:** Add `flash undeploy` command for endpoint management ([#121](https://github.com/runpod/tetra-rp/issues/121)) ([cd32ffc](https://github.com/runpod/tetra-rp/commit/cd32ffc40ac26c2f1aaa1235b044798ec7b9f605))
+
+
+### Code Refactoring
+
+* move endpoint deletion logic to proper abstraction layers ([#124](https://github.com/runpod/tetra-rp/issues/124)) ([c253d3b](https://github.com/runpod/tetra-rp/commit/c253d3b6f959dc8514b91b5e79f6f73fe9593b89))
+
+## [0.17.1](https://github.com/runpod/tetra-rp/compare/v0.17.0...v0.17.1) (2025-11-19)
+
+
+### Documentation
+
+* update readme for alpha release ([#120](https://github.com/runpod/tetra-rp/issues/120)) ([e8093d8](https://github.com/runpod/tetra-rp/commit/e8093d8ac878c13e578dc0ded823dfe8c73120f3))
+
+## [0.17.0](https://github.com/runpod/tetra-rp/compare/v0.16.1...v0.17.0) (2025-11-19)
+
+
+### Features
+
+* add user-friendly error messages for missing RUNPOD_API_KEY ([#117](https://github.com/runpod/tetra-rp/issues/117)) ([32dc093](https://github.com/runpod/tetra-rp/commit/32dc0937dda02a8b6178f3f5b2219f18de1f933e))
+
+
+### Documentation
+
+* improved flash init skeleton, contributing, and misc cleanup ([#118](https://github.com/runpod/tetra-rp/issues/118)) ([91acf1a](https://github.com/runpod/tetra-rp/commit/91acf1a87b5cdcf7f66c2bfa976f6132a8ca3cea))
+
 ## [0.16.1](https://github.com/runpod/tetra-rp/compare/v0.16.0...v0.16.1) (2025-11-14)